+++ ja.js
@@ -1213,6 +1213,10 @@ export default {
     other {カード}
   }`,
   'payment.options.default': 'デフォルト',
+  // Added 2020-05-27
+  'payment.method.add': 'Add a new payment method',
+  'payment.methods.adding': 'You new payment method has been added to your account.',
+  'payment.methods.delete': 'Delete payment method',
 }
 
 /* eslint-enable max-len */